flask/requirements.txt
Flask
Flask-PyMongo
Flask-JWT-Extended
python-dotenv
pymongo
